Override a FAILURE global status with ERROR
The shortcut test suite result method to add a test case result is
updated so that the ERROR test status is made prioritary over the
FAILURE one: initially, when a test ends in FAILURE but the global
status was set to FAILURE, this value was left. Now, this value is
overridden with ERROR, because an error during a test execution is
considered a "worse" condition that the test failing.