Ensure that inline source root is XSL stylesheet
The retrieval of the child element of inline source is extracted into a
new method, because it is now much more complex: the type of the object
is now checked against String and Element, and in the latter's case its
namespace URI and element name are compared to the standard XSL URI and
stylesheet, respectively. If any of these checks fails, the method
throws an error.