Apply fixes to the playlists stylesheet
The function parameter definitions no longer have a select attribute, as
it is not applicable to this element.
The context node is explicitly passed to the time-to-seconds function as
parameter, because if it is not given there is no context node for the
function to use (although I thought that was what the /x/path/func()
syntax enabled, but manifestly not).
Variable references in the seconds-to-time function are expressed as
such: prefixed with a dollar sign.